Done SpaceX more info for CardViews

+++ app/src/main/java/ru/myitschool/nasa_bootcamp/data/dto/spaceX/launches/SecondStage.kt
@@ -11,7 +11,7 @@ data class Payloads(
     @field:SerializedName("reused") val reused: Boolean,
     @field:SerializedName("nationality") val nationality: String,
     @field:SerializedName("manufacturer") val manufacturer: String,
-    @field:SerializedName("payload_mass_kg") val payload_mass_kg: Int,
+    @field:SerializedName("payload_mass_kg") val payload_mass_kg: Double,
     @field:SerializedName("payload_type") val payload_type: String,
     @field:SerializedName("reference_system") val reference_system: String
     )
